+++ .storybook/preview.ts
@@ -2,7 +2,7 @@ import { INITIAL_VIEWPORTS } from "@storybook/addon-viewport";
 import { type Preview } from "@storybook/react";
 import { initialize, mswLoader } from "msw-storybook-addon";
 
-import "../src/app/index.css";
+import "../src/index.css";
 
 // Initialize MSW
 initialize();
