+++ .storybook/test-runner.ts
@@ -2,10 +2,10 @@ import { type Page } from "playwright";
 import { injectAxe, checkA11y } from "axe-playwright";
 
 export default {
-	async preVisit(page: Page): Promise<void> {
+	async preVisit (page: Page): Promise<void> {
 		await injectAxe(page);
 	},
-	async postVisit(page: Page): Promise<void> {
+	async postVisit (page: Page): Promise<void> {
 		await checkA11y(page, "#storybook-root", {
 			detailedReport: true,
 			detailedReportOptions: {
