+++ package.json
@@ -103,6 +103,7 @@
     "cypress": "^13.15.0",
     "eslint-plugin-storybook": "^0.9.0",
     "jsdom": "^25.0.1",
+    "msw": "^2.4.13",
     "msw-storybook-addon": "^2.0.3",
     "playwright": "^1.48.1",
     "prop-types": "^15.8.1",
