+++ src/inbox/inbox.story.tsx
@@ -1,111 +1,104 @@
 import { expect, userEvent, findByRole, within } from "@storybook/test";
-import { http, HttpResponse } from 'msw';
+import { http, HttpResponse } from "msw";
 
-import InboxScreen from "./inbox";
-import { Default as TaskListDefault } from "../task-list/task-list.story";
-
+import { Default as TaskListDefault } from "../tasks/task-list.story";
+import { Inbox } from "./inbox";
 
 export default {
-  component: InboxScreen,
-  title: "InboxScreen",
+	component: Inbox,
+	title: "Inbox",
 };
 
 export const Default = {
-  parameters: {
-    msw: {
-      handlers: [
-        http.get("/tasks", () => {
-          return HttpResponse.json(TaskListDefault.args);
-        }),
-      ],
-    },
-  },
+	parameters: {
+		msw: {
+			handlers: [
+				http.get("/tasks", () => HttpResponse.json(TaskListDefault.args)),
+			],
+		},
+	},
 };
 
 export const Error = {
-  args: {
-    error: "Something",
-  },
-  parameters: {
-    msw: {
-      handlers: [
-        http.get("/tasks", () => {
-          return HttpResponse.json([]);
-        }),
-      ],
-    },
-  },
+	args: {
+		error: "Something",
+	},
+	parameters: {
+		msw: {
+			handlers: [
+				http.get("/tasks", () => HttpResponse.json([])),
+			],
+		},
+	},
 };
 
 export const PinTask = {
-  parameters: {
-    ...Default.parameters,
-  },
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-    const getTask = (id) => canvas.findByRole("listitem", { name: id });
+	parameters: {
+		...Default.parameters,
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		const getTask = (id) => canvas.findByRole("listitem", { name: id });
 
-    const itemToPin = await getTask("task-4");
-    // Find the pin button
-    const pinButton = await findByRole(itemToPin, "button", { name: "pin" });
-    // Click the pin button
-    await userEvent.click(pinButton);
-    // Check that the pin button is now a unpin button
-    const unpinButton = within(itemToPin).getByRole("button", {
-      name: "unpin",
-    });
-    await expect(unpinButton).toBeInTheDocument();
-  },
+		const itemToPin = await getTask("task-4");
+		// Find the pin button
+		const pinButton = await findByRole(itemToPin, "button", { name: "pin" });
+		// Click the pin button
+		await userEvent.click(pinButton);
+		// Check that the pin button is now a unpin button
+		const unpinButton = within(itemToPin).getByRole("button", {
+			name: "unpin",
+		});
+		await expect(unpinButton).toBeInTheDocument();
+	},
 };
 
 export const ArchiveTask = {
-  parameters: {
-    ...Default.parameters,
-  },
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-    const getTask = (id) => canvas.findByRole("listitem", { name: id });
+	parameters: {
+		...Default.parameters,
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		const getTask = (id) => canvas.findByRole("listitem", { name: id });
 
-    const itemToArchive = await getTask("task-2");
-    const archiveButton = await findByRole(itemToArchive, "button", {
-      name: "archiveButton-2",
-    });
-    await userEvent.click(archiveButton);
-  },
+		const itemToArchive = await getTask("task-2");
+		const archiveButton = await findByRole(itemToArchive, "button", {
+			name: "archiveButton-2",
+		});
+		await userEvent.click(archiveButton);
+	},
 };
 
 export const EditTask = {
-  parameters: {
-    ...Default.parameters,
-  },
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-    const getTask = (id) => canvas.findByRole("listitem", { name: id });
+	parameters: {
+		...Default.parameters,
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		const getTask = (id) => canvas.findByRole("listitem", { name: id });
 
-    const itemToEdit = await getTask("task-5");
-    const taskInput = await findByRole(itemToEdit, "textbox");
-    await userEvent.type(taskInput, " and disabled state");
-    await expect(taskInput.value).toBe(
-      "Fix bug in input error state and disabled state"
-    );
-  },
+		const itemToEdit = await getTask("task-5");
+		const taskInput = await findByRole(itemToEdit, "textbox");
+		await userEvent.type(taskInput, " and disabled state");
+		await expect(taskInput.value).toBe("Fix bug in input error state and disabled state");
+	},
 };
 
 export const DeleteTask = {
-  parameters: {
-    ...Default.parameters,
-  },
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-    const getTask = (id) => canvas.findByRole("listitem", { name: id });
+	parameters: {
+		...Default.parameters,
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		const getTask = (id) => canvas.findByRole("listitem", { name: id });
 
-    const itemToDelete = await getTask("task-1");
-    const deleteButton = await findByRole(itemToDelete, "button", {
-      name: "delete",
-    });
+		const itemToDelete = await getTask("task-1");
+		const deleteButton = await findByRole(itemToDelete, "button", {
+			name: "delete",
+		});
 
-    // Click the pin button
-    await userEvent.click(deleteButton);
-    await expect(canvas.getAllByRole("listitem").length).toBe(5);
-  },
+		// Click the pin button
+		await userEvent.click(deleteButton);
+		await expect(canvas.getAllByRole("listitem").length).toBe(5);
+	},
 };
