+++ src/login/login.css
@@ -1,99 +1,99 @@
 .login-form-container {
-  background: white;
-  padding-top: 1.5rem;
-  padding-bottom: 1.5rem;
-  padding-inline-start: 1.25rem;
-  padding-inline-end: 1.25rem;
+	background: white;
+	padding-top: 1.5rem;
+	padding-bottom: 1.5rem;
+	padding-inline-start: 1.25rem;
+	padding-inline-end: 1.25rem;
 }
 
 .login-form-container .login-form-wrapper {
-  display: flex;
+	display: flex;
 }
 .login-form-container .login-form-group {
-  width: 100%;
-  position: relative;
+	width: 100%;
+	position: relative;
 }
 
 .login-form-container .login-form-group .login-form-label {
-  display: block;
-  text-align: start;
-  font-size: 1rem;
-  margin-inline-end: 0.75rem;
-  margin-bottom: 0.5rem;
-  font-weight: normal;
+	display: block;
+	text-align: start;
+	font-size: 1rem;
+	margin-inline-end: 0.75rem;
+	margin-bottom: 0.5rem;
+	font-weight: normal;
 }
 .login-form-container .login-form-group .login-form-input {
-  width: 100%;
-  min-width: 0;
-  outline: none;
-  position: relative;
-  appearance: none;
-  font-size: 1rem;
-  padding-inline-start: 1rem;
-  padding-inline-end: 1rem;
-  height: 2.5rem;
-  border-radius: 0.375rem;
-  border: 1px solid;
-  border-color: inherit;
-  background: inherit;
+	width: 100%;
+	min-width: 0;
+	outline: none;
+	position: relative;
+	appearance: none;
+	font-size: 1rem;
+	padding-inline-start: 1rem;
+	padding-inline-end: 1rem;
+	height: 2.5rem;
+	border-radius: 0.375rem;
+	border: 1px solid;
+	border-color: inherit;
+	background: inherit;
 }
 
 .submit-button {
-  display: inline-flex;
-  appearance: none;
-  align-items: center;
-  justify-content: center;
-  user-select: none;
-  position: relative;
-  white-space: nowrap;
-  vertical-align: middle;
-  outline: none;
-  width: 100%;
-  line-height: 1.2rem;
-  border-radius: 0.375rem;
-  font-weight: 600;
-  height: 3rem;
-  min-height: 3rem;
-  font-size: 1rem;
-  padding-inline-end: 1.5rem;
-  padding-inline-start: 1.5rem;
-  color: white;
-  background-color: #1c3f53;
-  margin: 24px 0 0;
-  padding: 0 24px;
+	display: inline-flex;
+	appearance: none;
+	align-items: center;
+	justify-content: center;
+	user-select: none;
+	position: relative;
+	white-space: nowrap;
+	vertical-align: middle;
+	outline: none;
+	width: 100%;
+	line-height: 1.2rem;
+	border-radius: 0.375rem;
+	font-weight: 600;
+	height: 3rem;
+	min-height: 3rem;
+	font-size: 1rem;
+	padding-inline-end: 1.5rem;
+	padding-inline-start: 1.5rem;
+	color: white;
+	background-color: #1c3f53;
+	margin: 24px 0 0;
+	padding: 0 24px;
 }
 
 .loginscreen {
-  height: 100vh;
-  display: flex;
-  align-items: center;
+	height: 100vh;
+	display: flex;
+	align-items: center;
 }
 
 .login-screen-container {
-  width: 100%;
-  max-width: 28rem;
-  margin-inline-start: auto;
-  margin-inline-end: auto;
-  border: 1px solid #e2e8f0;
+	width: 100%;
+	max-width: 28rem;
+	margin-inline-start: auto;
+	margin-inline-end: auto;
+	border: 1px solid #e2e8f0;
 }
 
 .loginscreen-header {
-  background-color: #d3edf4;
-  padding-top: 1.5rem;
-  padding-bottom: 1.5rem;
-  padding-inline-start: 1.25rem;
-  padding-inline-end: 1.25rem;
+	background-color: #d3edf4;
+	padding-top: 1.5rem;
+	padding-bottom: 1.5rem;
+	padding-inline-start: 1.25rem;
+	padding-inline-end: 1.25rem;
 }
 
 .loginscreen-heading {
-  font-weight: bold;
-  font-size: 1.5rem;
-  line-height: 2rem;
-  color: #1c3f53;
+	font-weight: bold;
+	font-size: 1.5rem;
+	line-height: 2rem;
+	color: #1c3f53;
 }
 
 .loginscreen-text {
-  color: #4a5568;
-  font-size: 0.875rem;
-  line-height: 1.25rem;
+	color: #4a5568;
+	font-size: 0.875rem;
+	line-height: 1.25rem;
 }
